// src/domain/enums/permissions.ts
export enum Permission {
  EMPLOYEE_CREATE = "EMPLOYEE_CREATE",
  EMPLOYEE_UPDATE = "EMPLOYEE_UPDATE",
  EMPLOYEE_DELETE = "EMPLOYEE_DELETE",
  EMPLOYEE_VIEW = "EMPLOYEE_VIEW",

  PATIENT_CREATE = "PATIENT_CREATE",
  PATIENT_UPDATE = "PATIENT_UPDATE",
  PATIENT_DELETE = "PATIENT_DELETE",
  PATIENT_VIEW = "PATIENT_VIEW",

  SERVICE_CREATE = "SERVICE_CREATE",
  SERVICE_UPDATE = "SERVICE_UPDATE",
  SERVICE_DELETE = "SERVICE_DELETE",
  SERVICE_VIEW = "SERVICE_VIEW",

  APPOINTMENT_CREATE = "APPOINTMENT_CREATE",
  APPOINTMENT_UPDATE = "APPOINTMENT_UPDATE",
  APPOINTMENT_DELETE = "APPOINTMENT_DELETE",
  APPOINTMENT_VIEW = "APPOINTMENT_VIEW",

  BED_CREATE = "BED_CREATE",
  BED_UPDATE = "BED_UPDATE",
  BED_VIEW = "BED_VIEW",

  AUDIT_VIEW = "AUDIT_VIEW",
}